# swagger_api_desafiobdc.json
{
   "swagger":"2.0",
   "info":{
      "description":"API REST Spring Boot - Desafio BDC Company",
      "version":"1.0.0",
      "title":"API Seleção Java",
      "contact":{
         "name":"Wellesson Vieira",
         "url":"https://github.com/wellesson",
         "email":"wellesson15.cz@gmail.com"
      }
   },
   "host":"localhost:8080",
   "basePath":"/",
   "tags":[
      {
         "name":"cidade-controller",
         "description":"Cidade Controller"
      },
      {
         "name":"estado-controller",
         "description":"Estado Controller"
      },
      {
         "name":"perfil-controller",
         "description":"Perfil Controller"
      },
      {
         "name":"usuario-controller",
         "description":"Usuario Controller"
      }
   ],
   "paths":{
      "/api/cidade":{
         "get":{
            "tags":[
               "perfil-controller"
            ],
            "summary":"listar",
            "operationId":"listarUsingGET_2",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Cidade"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "perfil-controller"
            ],
            "summary":"salvar",
            "operationId":"salvarUsingPOST_2",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entidade a ser cadastrada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Cidade"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Cidade"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/cidade/{id}":{
         "get":{
            "tags":[
               "perfil-controller"
            ],
            "summary":"buscar",
            "operationId":"buscarUsingGET_2",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Cidade"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "perfil-controller"
            ],
            "summary":"alterar",
            "operationId":"alterarUsingPUT_2",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entida a ser atualizada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Cidade"
                  }
               },
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Cidade"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "perfil-controller"
            ],
            "summary":"remover",
            "operationId":"removerUsingDELETE_2",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/api/estado":{
         "get":{
            "tags":[
               "estado-controller"
            ],
            "summary":"listar",
            "operationId":"listarUsingGET_1",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/UnidadeFederativa"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "estado-controller"
            ],
            "summary":"salvar",
            "operationId":"salvarUsingPOST_1",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entidade a ser cadastrada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/UnidadeFederativa"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/UnidadeFederativa"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/estado/{id}":{
         "get":{
            "tags":[
               "estado-controller"
            ],
            "summary":"buscar",
            "operationId":"buscarUsingGET_1",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/UnidadeFederativa"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "estado-controller"
            ],
            "summary":"alterar",
            "operationId":"alterarUsingPUT_1",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entida a ser atualizada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/UnidadeFederativa"
                  }
               },
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/UnidadeFederativa"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "estado-controller"
            ],
            "summary":"remover",
            "operationId":"removerUsingDELETE_1",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/api/perfil":{
         "get":{
            "tags":[
               "cidade-controller"
            ],
            "summary":"listar",
            "operationId":"listarUsingGET",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/Perfil"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "cidade-controller"
            ],
            "summary":"salvar",
            "operationId":"salvarUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entidade a ser cadastrada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Perfil"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Perfil"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/perfil/{id}":{
         "get":{
            "tags":[
               "cidade-controller"
            ],
            "summary":"buscar",
            "operationId":"buscarUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Perfil"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "cidade-controller"
            ],
            "summary":"alterar",
            "operationId":"alterarUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"entity",
                  "description":"Objeto entida a ser atualizada.",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Perfil"
                  }
               },
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/Perfil"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "cidade-controller"
            ],
            "summary":"remover",
            "operationId":"removerUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK"
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      },
      "/api/usuario":{
         "get":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Listar todos os usuários por UF",
            "operationId":"listarUsingGET_3",
            "produces":[
               "*/*"
            ],
            "responses":{
               "200":{
                  "description":"Consulta realizada com sucesso.",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/ResponseUsuarioDTO"
                     }
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Cadastra usuário",
            "operationId":"incluirUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"requestUsuarioDTO",
                  "description":"requestUsuarioDTO",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/RequestUsuarioDTO"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Cadastro realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/usuario/perfil/{codigoPerfil}":{
         "get":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Listar todos os usuários por Perfil",
            "operationId":"consultarPorPerfilUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"codigoPerfil",
                  "in":"path",
                  "description":"codigoPerfil",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"Consulta realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/usuario/uf/{codigoUf}":{
         "get":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Listar todos os usuários por UF",
            "operationId":"consultarPorUfUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"codigoUf",
                  "in":"path",
                  "description":"codigoUf",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"Consulta realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         }
      },
      "/api/usuario/{id}":{
         "get":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Busca usuário por ID",
            "operationId":"consultarPorIdUsingGET",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"Consulta realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Atualizar usuário",
            "operationId":"atualizarUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               },
               {
                  "in":"body",
                  "name":"requestUsuarioDTO",
                  "description":"requestUsuarioDTO",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/RequestUsuarioDTO"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Atualização realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "usuario-controller"
            ],
            "summary":"Exclui usuário por ID",
            "operationId":"excluirUsingDELETE",
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"Exclusão realizada com sucesso.",
                  "schema":{
                     "$ref":"#/definitions/ResponseUsuarioDTO"
                  }
               },
               "204":{
                  "description":"No Content"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               }
            },
            "deprecated":false
         }
      }
   },
   "definitions":{
      "Cidade":{
         "type":"object",
         "properties":{
            "estado":{
               "$ref":"#/definitions/UnidadeFederativa"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "nome":{
               "type":"string"
            }
         },
         "title":"Cidade"
      },
      "Endereco":{
         "type":"object",
         "properties":{
            "bairro":{
               "type":"string"
            },
            "cidade":{
               "$ref":"#/definitions/Cidade"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "logradouro":{
               "type":"string"
            },
            "numero":{
               "type":"string"
            }
         },
         "title":"Endereco"
      },
      "Perfil":{
         "type":"object",
         "properties":{
            "descricao":{
               "type":"string"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            }
         },
         "title":"Perfil"
      },
      "RequestEnderecoDTO":{
         "type":"object",
         "properties":{
            "bairro":{
               "type":"string"
            },
            "codigoCidade":{
               "type":"integer",
               "format":"int64"
            },
            "logradouro":{
               "type":"string"
            },
            "numero":{
               "type":"string"
            }
         },
         "title":"RequestEnderecoDTO"
      },
      "RequestUsuarioDTO":{
         "type":"object",
         "properties":{
            "codigoPerfil":{
               "type":"integer",
               "format":"int64"
            },
            "cpf":{
               "type":"string"
            },
            "endereco":{
               "$ref":"#/definitions/RequestEnderecoDTO"
            },
            "nome":{
               "type":"string"
            },
            "senha":{
               "type":"string"
            }
         },
         "title":"RequestUsuarioDTO"
      },
      "ResponseCidadeDTO":{
         "type":"object",
         "properties":{
            "estado":{
               "$ref":"#/definitions/ResponseEstadoDTO"
            },
            "nome":{
               "type":"string"
            }
         },
         "title":"ResponseCidadeDTO"
      },
      "ResponseEnderecoDTO":{
         "type":"object",
         "properties":{
            "bairro":{
               "type":"string"
            },
            "cidade":{
               "$ref":"#/definitions/ResponseCidadeDTO"
            },
            "logradouro":{
               "type":"string"
            },
            "numero":{
               "type":"string"
            }
         },
         "title":"ResponseEnderecoDTO"
      },
      "ResponseEstadoDTO":{
         "type":"object",
         "properties":{
            "nome":{
               "type":"string"
            },
            "sigla":{
               "type":"string"
            }
         },
         "title":"ResponseEstadoDTO"
      },
      "ResponsePerfilDTO":{
         "type":"object",
         "properties":{
            "descricao":{
               "type":"string"
            }
         },
         "title":"ResponsePerfilDTO"
      },
      "ResponseUsuarioDTO":{
         "type":"object",
         "properties":{
            "cpf":{
               "type":"string"
            },
            "endereco":{
               "$ref":"#/definitions/ResponseEnderecoDTO"
            },
            "nome":{
               "type":"string"
            },
            "perfil":{
               "$ref":"#/definitions/ResponsePerfilDTO"
            },
            "senha":{
               "type":"string"
            }
         },
         "title":"ResponseUsuarioDTO"
      },
      "UnidadeFederativa":{
         "type":"object",
         "properties":{
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "nome":{
               "type":"string"
            },
            "sigla":{
               "type":"string"
            }
         },
         "title":"UnidadeFederativa"
      },
      "Usuario":{
         "type":"object",
         "properties":{
            "cpf":{
               "type":"string"
            },
            "endereco":{
               "$ref":"#/definitions/Endereco"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "nome":{
               "type":"string"
            },
            "perfil":{
               "$ref":"#/definitions/Perfil"
            },
            "senha":{
               "type":"string"
            }
         },
         "title":"Usuario"
      }
   }
}